refactor: AdminBanner 구조 변경
