+++ admin/src/main/java/com/kernel/admin/domain/enums/BannerStatus.java
@@ -1,4 +1,4 @@
-package com.kernel.admin.domain.enumerate;
+package com.kernel.admin.domain.enums;
 
 public enum BannerStatus {
     ACTIVE("활성"),
